pyproject.toml
[tool.poetry]
name = "n5project"
version = "0.1.0"
description = "N5 Project"
authors = ["Hugo Silva <hugofer93@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
python-dotenv = "~0.20"
Django = "~3.2"
psycopg2 = "~2.9"
uWSGI = "~2.0"
djangorestframework = "~3.13"
drf-spectacular = "~0.22"
drf-ujson2 = "~1.6"
djangorestframework-camel-case = "~1.3"
django-colorfield = "~0.8"
djangorestframework-simplejwt = "~5.2"

[tool.poetry.dev-dependencies]
django-silk = "~4.3"

[build-system]
